Change score type from u32 to f64 in score macro

+++ crates/autograder-test/src/lib.rs
@@ -8,7 +8,7 @@ use std::time::Duration;
 #[macro_export]
 macro_rules! score {
     ($score:expr) => {
-        let score: u32 = $score;
+        let score: f64 = $score;
         println!("autograder: score={score}");
     };
 }
